+++ src/common/src/hash/vnode.rs
@@ -19,7 +19,6 @@ use crate::hash::HashCode;
 /// Parallel unit is the minimal scheduling unit.
 // TODO: make it a newtype
 pub type ParallelUnitId = u32;
-pub type VnodeMapping = Vec<ParallelUnitId>;
 
 /// `VirtualNode` (a.k.a. VNode) is a minimal partition that a set of keys belong to. It is used for
 /// consistent hashing.
